refactor: dto req/res 통합
