--- booker-server/src/main/java/com/bookerapp/core/presentation/dto/event/UpdateEventRequest.java
@@ -1,15 +1,0 @@
-package com.bookerapp.core.presentation.dto.event;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor
-public class UpdateEventRequest {
-    private String title;
-    private String description;
-    private LocalDateTime startTime;
-    private LocalDateTime endTime;
-}
